requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-countries==7.3.2
django-environ==0.8.1
django-paypal==2.0
idna==3.3
mysqlclient==2.1.0
orjson==3.7.2
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
